+++ @Shared/General/Attributes.cs
@@ -1,0 +1,35 @@
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.IO;
+using System.Linq;
+using System.Management;
+using System.Net;
+using System.Net.Http;
+using System.Reflection;
+using System.Runtime.InteropServices;
+using System.Security.Principal;
+using System.Text;
+using System.Threading.Tasks;
+using System.Xml.Linq;
+
+namespace Shared.Attributes
+{
+    /// <summary>
+    /// Parameter attribute. Given parameter can be null.
+    /// </summary>
+    [AttributeUsage(AttributeTargets.Parameter, AllowMultiple = false, Inherited = false)]
+    public class MaybeNull : Attribute
+    {
+
+    }
+
+    /// <summary>
+    /// Parameter attribute. Given parameter must not be null.
+    /// </summary>
+    [AttributeUsage(AttributeTargets.Parameter, AllowMultiple = false, Inherited = false)]
+    public class NeverNull : Attribute
+    {
+
+    }
+}