+++ @Shared/General/Logger.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Runtime.CompilerServices;
 
-namespace Shared
+namespace Shared.Loggers
 {
     /// <summary>
     /// Methods to print to standard output.
@@ -14,8 +14,8 @@ namespace Shared
         [System.Diagnostics.Conditional("DEBUG")]
         public static void PrintDebug(string msg, [CallerMemberName] string caller = "")
         {
-            Console.Write($"[DEBUG] [{caller}] ");
-            Console.WriteLine(msg);
+            System.Console.Write($"[DEBUG] [{caller}] ");
+            System.Console.WriteLine(msg);
         }
 
         /// <summary>
@@ -23,8 +23,8 @@ namespace Shared
         /// </summary>
         public static void Print(string msg)
         {
-            Console.Write($"[INFO] ");
-            Console.WriteLine(msg);
+            System.Console.Write($"[INFO] ");
+            System.Console.WriteLine(msg);
         }
 
         /// <summary>
@@ -32,8 +32,8 @@ namespace Shared
         /// </summary>
         public static void PrintWarning(string msg, [CallerMemberName] string caller = "")
         {
-            Console.Write($"[WARNING] [{caller}] ");
-            Console.WriteLine(msg);
+            System.Console.Write($"[WARNING] [{caller}] ");
+            System.Console.WriteLine(msg);
         }
 
         /// <summary>
@@ -41,8 +41,8 @@ namespace Shared
         /// </summary>
         public static void PrintError(string msg, [CallerMemberName] string caller = "")
         {
-            Console.Write($"[Exception/Error] [{caller}] ");
-            Console.WriteLine(msg);
+            System.Console.Write($"[Exception/Error] [{caller}] ");
+            System.Console.WriteLine(msg);
         }
 
         /// <summary>
@@ -50,7 +50,7 @@ namespace Shared
         /// </summary>
         public static void PrintException(Exception exception)
         {
-            Console.WriteLine(exception.ToString());
+            System.Console.WriteLine(exception.ToString());
         }
     }
 }