+++ @Shared/UnityMod/Settings.cs
@@ -1,4 +1,5 @@
 ﻿using Newtonsoft.Json;
+using Shared.Loggers;
 using System;
 using System.Collections.Generic;
 using System.IO;